[android] - add GLSurfaceView RecylerView example to test for regressions
