+++ platform/android/MapboxGLAndroidSDKTestApp/src/main/res/layout/item_map_gl.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.mapbox.mapboxsdk.maps.MapView
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        android:id="@id/mapView"
+        android:layout_width="match_parent"
+        android:layout_height="256dp"
+        app:mapbox_cameraTargetLat="45.38301927899065"
+        app:mapbox_cameraTargetLng="8.63525390625"
+        app:mapbox_cameraZoom="7"/>
