+++ platform/android/MapboxGLAndroidSDKTestApp/src/main/res/layout/item_map_texture.xml
